web: Move some internal files to internal/
