+++ web/packages/core/src/internal/i18n.ts
