+++ web/packages/core/src/internal/register-element.ts
